+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/services/pin_code_service.dart
@@ -11,9 +11,8 @@ abstract class PinCodeService {
   /// the length should be less than 10 digits
   Future<int> getPinLength();
 
-  /// Encrypts the string and stores it in the device secure storage. It has a
-  /// add a default implementation to return what is the input. Returns the
-  /// encrypted pinCode from the device secure storage.
+  /// Encrypts the provided pin code and returns the encrypted string. It has a
+  /// default implementation to return what is the input. 
   Future<String> encryptPinCode(String pinCode) => Future.value(pinCode);
 
   /// Returns the pin code or null if it is not saved in the device memory yet.